+++ Mcs3Rob/RobParameter.cs
@@ -4,12 +4,12 @@ namespace Mcs3Rob
     {
         public RobElementType ElementType { get; set; }
 
-        public int? Bitmask { get; set; }
+        public uint? Bitmask { get; set; }
         public RobDirection? Direction { get; set; }
 
-        public int MinimumLimit { get; set; }
-        public int MaximumLimit { get; set; }
-        public int Address { get; set; }
+        public long MinimumLimit { get; set; }
+        public long MaximumLimit { get; set; }
+        public uint Address { get; set; }
         public RobDisplayMode DisplayMode { get; set; }
         public string Description { get; set; }
         public string VariableName { get; set; }
